requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow
apache-airflow-providers-http==4.12.0
apache-airflow-providers-google==10.22.0
requests
pandas